+++ StorageManage/StorageManage/ButtonClick/DeleteDetail.cs
@@ -25,12 +25,15 @@ namespace StorageManage.ButtonClick
             DataRow DR = DRV.Row;
             object[] arr = DR.ItemArray;
             MySqlDataReader reader = window.ex.returnResult("select idorders from orders where iddetails="+arr[0]);
+            if (reader == null) { return; }
             if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
             window.ex.closeCon();
             reader = window.ex.returnResult("select recordid from devices_details where iddetails=" + arr[0]);
+            if (reader == null) { return; }
             if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
             window.ex.closeCon();
             reader = window.ex.returnResult("select recordid from malfunctions_details where iddetails=" + arr[0]);
+            if (reader == null) { return; }
             if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
             window.ex.closeCon();
             window.ex.ExecuteWithoutRedaer("delete from details where iddetails=" + arr[0]);
