+++ StorageManage/StorageManage/ButtonClick/DeleteTypeOfDevice.cs
@@ -25,9 +25,11 @@ namespace StorageManage.ButtonClick
             DataRow DR = DRV.Row;
             object[] arr = DR.ItemArray;
             MySqlDataReader reader = window.ex.returnResult("select iddevices from devices where idtypes=" + arr[0]);
+            if (reader == null) { return; }
             if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
             window.ex.closeCon();
             reader = window.ex.returnResult("select idmalfunctions from malfunctions where idtypes=" + arr[0]);
+            if (reader == null) { return; }
             if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
             window.ex.closeCon();
             window.ex.ExecuteWithoutRedaer("delete from typeofdevices where idtypes=" + arr[0]);
