+++ StorageManage/StorageManage/ButtonClick/DeleteUser.cs
@@ -25,6 +25,7 @@ namespace StorageManage.ButtonClick
             DataRow DR = DRV.Row;
             object[] arr = DR.ItemArray;
             MySqlDataReader reader = window.ex.returnResult("select id from orders where id=" + arr[0]);
+            if (reader == null) { return; }
             if (reader.HasRows) { window.ex.closeCon(); MessageBox.Show("Невозможно удалить запись"); return; }
             window.ex.closeCon();
             window.ex.ExecuteWithoutRedaer("delete from users where id=" + arr[0]);
